+++ src/graph/memory_backing/disk_backing/disk_mut_refs.rs
@@ -53,7 +53,7 @@ where
         let _ = disk_vec.set(self.index, &node);
     }
 
-    fn set_count(self, count: u16) {
+    fn set_count(self, count: usize) {
         let mut disk_vec = self.disk_vec.borrow_mut();
         let mut node = disk_vec.get(self.index).unwrap();
         node.weight.set_count(count);
